Refactor ATLS implementation: remove old atls module, add new grpc/atls functionality, and update related configurations
